pyproject.toml
[project]
name = "report-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "argparse>=1.4.0",
    "openai>=1.0.0",
    "google-genai>=1.0.0",
    "anthropic>=0.3.0",
    "tiktoken>=0.5.0",
]

[dependency-groups]
dev = [
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pytest>=7.0.0",
]
